chore: restructure build files
Simplify tsconfig setup:
- drop unused composite (no project-ref builds, all noEmit)
- drop unused jsx preserve (Svelte not JSX)
- drop noImplicitAny override (use strict default)
- include tests/main and tests/renderer so test files type-check
- remove dead src/preload/*.d.ts include glob
- typecheck:node no longer needs --composite false workaround
