+++ e2e/playwright.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "@playwright/test";
 
 export default defineConfig({
-  testDir: "./e2e",
+  testDir: "./tests",
   testMatch: /.*\.spec\.ts$/,
   fullyParallel: false,
   workers: 1,
