+++ electron-builder.json
@@ -1,0 +1,46 @@
+{
+  "$schema": "https://raw.githubusercontent.com/electron-userland/electron-builder/master/packages/app-builder-lib/scheme.json",
+  "appId": "com.diodedj.app",
+  "productName": "DiodeDJ",
+  "directories": {
+    "output": "release"
+  },
+  "files": ["out/**/*", "package.json"],
+  "asarUnpack": [
+    "node_modules/ffmpeg-static/**",
+    "node_modules/better-sqlite3/**"
+  ],
+  "mac": {
+    "target": [
+      {
+        "target": "dmg",
+        "arch": ["arm64", "x64"]
+      },
+      {
+        "target": "zip",
+        "arch": ["arm64", "x64"]
+      }
+    ],
+    "category": "public.app-category.music",
+    "extendInfo": {
+      "NSDesktopFolderUsageDescription": "DiodeDJ scans audio files on your Desktop.",
+      "NSDocumentsFolderUsageDescription": "DiodeDJ scans audio files in Documents.",
+      "NSDownloadsFolderUsageDescription": "DiodeDJ scans audio files in Downloads.",
+      "NSMusicLibraryUsageDescription": "DiodeDJ scans your music library.",
+      "NSRemovableVolumesUsageDescription": "DiodeDJ scans audio on external drives.",
+      "NSNetworkVolumesUsageDescription": "DiodeDJ scans audio on network volumes."
+    }
+  },
+  "win": {
+    "target": [
+      {
+        "target": "nsis",
+        "arch": ["x64"]
+      }
+    ]
+  },
+  "linux": {
+    "target": ["AppImage", "deb"],
+    "category": "AudioVideo"
+  }
+}
