+++ tsconfig.node.json
@@ -1,25 +1,20 @@
 {
+  "$schema": "https://json.schemastore.org/tsconfig",
+  "extends": "./tsconfig.base.json",
   "compilerOptions": {
-    "target": "ES2022",
-    "module": "ESNext",
-    "moduleResolution": "Bundler",
-    "lib": ["ES2022"],
-    "types": ["node", "electron-vite/node"],
-    "strict": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "resolveJsonModule": true,
-    "noEmit": true
+    "types": ["node", "electron-vite/node"]
   },
   "include": [
     "src/main/**/*",
     "src/preload.ts",
     "src/types.ts",
+    "tests/main/**/*",
+    "e2e/",
+    "commitlint.config.ts",
     "electron.vite.config.ts",
     "eslint.config.mjs",
-    "vitest.config.ts",
-    "playwright.config.ts",
-    "commitlint.config.ts"
+    "prettier.config.mjs",
+    "svelte.config.mjs",
+    "vitest.config.ts"
   ]
 }
