+++ tsconfig.web.json
@@ -1,22 +1,9 @@
 {
+  "$schema": "https://json.schemastore.org/tsconfig",
+  "extends": "./tsconfig.base.json",
   "compilerOptions": {
-    "target": "ES2022",
-    "module": "ESNext",
-    "moduleResolution": "Bundler",
-    "lib": ["ES2022", "DOM", "DOM.Iterable"],
-    "types": ["svelte"],
-    "strict": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "resolveJsonModule": true,
-    "noEmit": true,
-    "allowJs": true,
-    "checkJs": false
+    "lib": ["ESNext", "DOM", "DOM.Iterable"],
+    "types": ["svelte"]
   },
-  "include": [
-    "src/renderer/**/*.ts",
-    "src/renderer/**/*.svelte.ts",
-    "src/types.ts"
-  ]
+  "include": ["src/renderer/**/*", "src/types.ts", "tests/renderer/**/*"]
 }
